Fixed server to set export port to listen port + 1.   Added subinspectors for looking at other individuals, but it is difficult to use in the subinspector pane. I've added a button that toggles a split-screen view, and it works better there.  A few other misc. fixes

+++ MedSavantClient/src/main/java/org/ut/biolab/medsavant/client/region/RegionListAggregatePanel.java
@@ -340,8 +340,8 @@ public class RegionListAggregatePanel extends AggregatePanel {
                         ReferenceController.getInstance().getCurrentReferenceID(),
                         FilterController.getInstance().getAllFilterConditions(),
                         reg.getChrom(),
-                        reg.getStart(),
-                        reg.getEnd());
+                        (int)reg.getStart(),
+                        (int)reg.getEnd());
                 updatePatientCount(reg, recordsInRegion);
                 if (Thread.interrupted()) {
                     throw new InterruptedException();
